+++ lib/route/src/Obelisk/Route.hs
@@ -47,6 +47,7 @@ module Obelisk.Route
   , tryDecode
   , hoistCheck
   , hoistParse
+  , generalizeIdentity
   , mapSome
   , rPrism
   , _R
@@ -112,6 +113,7 @@ module Obelisk.Route
   , obeliskRouteEncoder
   , obeliskRouteSegment
   , pageNameEncoder
+  , pathQueryEncoder
   , handleEncoder
   , someSumEncoder
   , Void1
@@ -366,6 +368,12 @@ hoistCheck f (Encoder x) = Encoder (f x)
 hoistParse :: (Functor check)
   => (forall t. parse t -> parse' t) -> Encoder check parse a b -> Encoder check parse' a b
 hoistParse f (Encoder x) = Encoder (fmap (\(EncoderImpl dec enc) -> EncoderImpl (f . dec) enc) x)
+
+generalizeIdentity
+  :: (Functor check, Applicative parse)
+  => Encoder check Identity a b
+  -> Encoder check parse a b
+generalizeIdentity = hoistParse (pure . runIdentity)
 
 -- | Check an 'Encoder', transforming it into one whose check monad is anything we want (usually Identity).
 checkEncoder :: (Applicative check', Functor check)
@@ -1108,6 +1116,14 @@ void1Encoder = Encoder $ pure $ EncoderImpl
 instance GShow Void1 where
   gshowsPrec _ = \case {}
 
+-- | Encode a 'PathQuery' as 'Text'
+pathQueryEncoder :: (Applicative check, Applicative parse) => Encoder check parse PathQuery Text
+pathQueryEncoder = unsafeMkEncoder $ EncoderImpl
+  { _encoderImpl_encode = \(k, v) -> T.pack $ k <> v
+  , _encoderImpl_decode = \r ->
+      pure $ bimap T.unpack T.unpack $ T.breakOn "?" r
+  }
+
 -- | Given a backend route and a checked route encoder, render the route (path
 -- and query string). See 'checkEncoder' for how to produce a checked encoder.
 renderBackendRoute
@@ -1132,9 +1148,9 @@ renderObeliskRoute
   -> R (FullRoute a b)
   -> Text
 renderObeliskRoute e r =
-  let enc :: Encoder Identity (Either Text) (R (FullRoute a b)) PathQuery
-      enc = (pageNameEncoder . hoistParse (pure . runIdentity) e)
-  in (T.pack . uncurry (<>)) $ encode enc r
+  let enc :: Encoder Identity (Either Text) (R (FullRoute a b)) Text
+      enc = (pathQueryEncoder . pageNameEncoder . generalizeIdentity e)
+  in encode enc r
 
 -- | As per the 'unsafeTshowEncoder' but does not use the 'Text' type.
 --